Add controller advisor for loan

+++ src/main/java/com/samadihadis/loanmanagementsystem/service/LoanService.java
@@ -1,19 +1,21 @@
 package com.samadihadis.loanmanagementsystem.service;
 
 
+import com.samadihadis.loanmanagementsystem.dto.loan.CreateLoanRequest;
 import com.samadihadis.loanmanagementsystem.dto.loan.LoanResponse;
 import com.samadihadis.loanmanagementsystem.entity.Customer;
 import com.samadihadis.loanmanagementsystem.entity.Loan;
 import com.samadihadis.loanmanagementsystem.enums.LoanStatus;
 import com.samadihadis.loanmanagementsystem.enums.LoanType;
+import com.samadihadis.loanmanagementsystem.exception.loan.InvalidLoanDataException;
+import com.samadihadis.loanmanagementsystem.exception.loan.LoanNotFoundException;
 import com.samadihadis.loanmanagementsystem.repository.CustomerRepository;
 import com.samadihadis.loanmanagementsystem.repository.LoanRepository;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -22,7 +24,31 @@ public class LoanService {
     private final LoanRepository loanRepository;
     public final CustomerRepository customerRepository;
 
-    public Loan createLoan(Loan loan, @NotNull Long customerId) {
+    public Loan createLoan(CreateLoanRequest request) {
+
+        if (request.getPrincipalAmount() == null || request.getPrincipalAmount().signum() <= 0) {
+            throw new InvalidLoanDataException("مبلغ وام باید مثبت و بیشتر از صفر باشد.");
+        }
+
+        if (request.getTerm() == null || request.getTerm() < 6 || request.getTerm() > 36) {
+            throw new InvalidLoanDataException("مدت وام باید بین ۶ تا ۳۶ ماه باشد.");
+        }
+
+        Customer customer = customerRepository.findById(request.getCustomerId())
+                .orElseThrow(() -> new LoanNotFoundException(
+                     String.format("مشتری با شناسه %d یافت نشد.", request.getCustomerId())
+                ));
+
+        Loan loan = new Loan();
+        loan.setPrincipalAmount(request.getPrincipalAmount());
+        loan.setInterestRate(request.getInterestRate());
+        loan.setTerm(request.getTerm());
+        loan.setStartDate(request.getStartDate());
+        loan.setLoanType(request.getLoanType());
+        loan.setCustomer(customer);
+
+        loan.setLoanStatus(LoanStatus.PENDING);
+
         return loanRepository.save(loan);
     }
 
@@ -31,28 +57,39 @@ public class LoanService {
     }
 
     public Loan getLoanById(Long id) {
-        Optional<Loan> loan = loanRepository.findById(id);
-        return loan.orElse(null);
+        return loanRepository.findById(id)
+                .orElseThrow(() -> new LoanNotFoundException(
+                        String.format("مشتری با شناسه %d یافت نشد.", id)
+                ));
     }
 
-    public void deleteLoan(long id) {
+    public void deleteLoan(Long id) {
+        if (!loanRepository.existsById(id)) {
+            throw new LoanNotFoundException(
+                    String.format("مشتری با شناسه %d یافت نشد.", id)
+            );
+        }
         loanRepository.deleteById(id);
     }
 
     public List<Loan> getLoanByCustomerId(Long customerId) {
 
         if (customerId == null) {
-            throw new IllegalArgumentException("Customer and customer ID must not be null");
+            throw new InvalidLoanDataException("شناسه مشتری نمی‌تواند خالی باشد");
         }
-        Customer customer = customerRepository.findById(customerId)
-                .orElseThrow(() -> new RuntimeException("برای مشتری با شناسه " + customerId + " وامی یافت نشد."));
+
+        customerRepository.findById(customerId)
+                .orElseThrow(() -> new LoanNotFoundException(
+                        String.format("مشتری با شناسه %d یافت نشد.", customerId)
+                ));
 
         List<Loan> loans = loanRepository.findByCustomerId(customerId);
 
         if (loans.isEmpty()) {
-            throw new RuntimeException("برای مشتری با شناسه " + customerId + " وامی یافت نشد.");
+            throw new LoanNotFoundException(
+                    String.format("مشتری با شناسه %d یافت نشد.", customerId)
+            );
         }
-
         return loans;
     }
 
@@ -67,39 +104,17 @@ public class LoanService {
     public List<Loan> getLoanByCustomerIdAndStatus(Long customerId, LoanStatus loanStatus) {
 
         if (customerId == null) {
-            throw new IllegalArgumentException("Customer and customer ID must not be null");
+            throw new InvalidLoanDataException("شناسه مشتری نمی‌تواند خالی باشد.");
         }
-
-        Customer customer = customerRepository.findById(customerId)
-                .orElseThrow(() -> new RuntimeException("برای مشتری با شناسه " + customerId + " وامی یافت نشد."));
-
-        List<Loan> loans = loanRepository.findByCustomerIdAndLoanStatus(customerId, loanStatus);
-
-        if (loans.isEmpty()) {
-            throw new IllegalArgumentException(String.format("برای مشتری با شناسه %d و همچنین وضعیت %s وامی یافت نشد" ,
-                    customerId , loanStatus));
-        }
-
-        return loans;
+        return loanRepository.findByCustomerIdAndLoanStatus(customerId, loanStatus);
     }
 
     public List<Loan> getLoanByCustomerIdAndType(Long customerId, LoanType loanType) {
 
         if (customerId == null) {
-            throw new IllegalArgumentException("Customer and customer ID must not be null");
+            throw new InvalidLoanDataException("شناسه مشتری نمی‌تواند خالی باشد.");
         }
-
-        Customer customer = customerRepository.findById(customerId)
-                .orElseThrow(() -> new RuntimeException("برای مشتری با شناسه " + customerId + " وامی یافت نشد."));
-
-        List<Loan> loans = loanRepository.findByCustomerIdAndLoanType(customerId, loanType);
-
-        if (loans.isEmpty()) {
-            throw new IllegalArgumentException(String.format("برای مشتری با شناسه %d و همچنین نوع %s وامی یافت نشد" ,
-                    customerId , loanType));
-        }
-
-        return loans;
+        return loanRepository.findByCustomerIdAndLoanType(customerId, loanType);
     }
 
     public LoanResponse toResponse(Loan loan) {
